Create PlayerDialog.py
PlayDialog.py opens a dialog box window that prompts the user to add a player to the selected team name (based on TeamPlayer.py).  The user can input the player's name, age, and position.  After clicking ok, a newly created Player namedtuple is created and appended to the list of players in the selected team (in TeamPlayer.py).